tagulous: make notifications dojo_async compatible
