+++ dojo/notifications/helper.py
@@ -7,10 +7,7 @@ from django.template.loader import render_to_string
 from django.db.models import Q, Count, Prefetch
 from django.urls import reverse
 from dojo.celery import app
-from django.db import models
-from django.forms.models import model_to_dict
-from dojo.decorators import we_want_async, dojo_async_task
-from django.db.models.query import QuerySet
+from dojo.decorators import we_want_async, dojo_async_task, convert_kwargs_if_async
 
 logger = logging.getLogger(__name__)
 
@@ -110,38 +107,6 @@ def create_notification_message(event, user, notification_type, *args, **kwargs)
             notification_message = render_to_string('notifications/other.tpl', kwargs)
 
     return notification_message if notification_message else ''
-
-
-def model_to_dict_with_tags(model):
-    converted = model_to_dict(model)
-    if 'tags' in converted:
-        # further conversion needed from Tag Queryset to strings
-        converted['tags'] = converted['tags'].values_list()
-    logger.debug('dict: %s', converted)
-    return converted
-
-
-def list_of_models_to_dict_with_tags(model_list):
-    result = []
-    for model in model_list:
-        result.append(model_to_dict_with_tags(model))
-    return result
-
-
-def convert_kwargs_if_async(**kwargs):
-    if we_want_async():
-        # not sync means using celery for notifications.
-        # sending full model instances to celery is bad practice.
-        # and any models with tags cannot be sent to celery due to serialization problems with celery
-        # we convert all model instances into dictionaries
-        for key, value in kwargs.items():
-            if isinstance(value, models.Model):
-                kwargs[key] = model_to_dict_with_tags(value)
-            elif isinstance(value, list):
-                kwargs[key] = list_of_models_to_dict_with_tags(value)
-            elif isinstance(value, QuerySet):
-                kwargs[key] = list_of_models_to_dict_with_tags(list(value))
-    return kwargs
 
 
 def process_notifications(event, notifications=None, **kwargs):
